Feat: Handle new form for enrichment
