+++ src/app/custom/translations.tsv
@@ -865,3 +865,19 @@
 "enrichment_see_logs"	"See logs"	"Voir les logs"
 "enrichment_status"	"Status"	"Statut"
 "enrichment_preview_description"	"Value from source column or from advanced rule"	"Valeur de la colonne source ou de la règle avancée"
+"enrichment_added_success"	"Enrichment added successfully"	"Enrichissement ajouté avec succès"
+"enrichment_updated_success"	"Enrichment updated successfully"	"Enrichissement modifié avec succès"
+"enrichment_preview_no_data"	"No data"	"Aucune donnée"	
+"enrichment_deleted_success"	"Enrichment deleted successfully"	"Enrichissement supprimé avec succès"
+"enrichment_status_pending"	"Pending"	"En attente"
+"enrichment_status_running"	"Running"	"En cours"
+"enrichment_status_paused"	"Paused"	"En pause"
+"enrichment_status_done"	"Done"	"Terminé"
+"enrichment_status_not_started"	"Not started"	"Non démarré"
+"enrichment_status_error"	"Failed"	"Échoué"
+"enrichment_logs_error" 	"Failed to get logs"	"Impossible de récupérer les logs"
+"metadata"	"Metadata"	"Métadonnées"
+"affiliation"	"Affiliation"	"Affiliation"
+"geography"	"Geography"	"Géographie"
+"classification"	"Classification"	"Classification"
+"indexation"	"Indexation"	"Indexation"
